src/requirements.txt
fastapi[all]
pydantic
firebase-admin
python-dotenv
pydantic_settings